requirements.txt
keyboard
dbus-python
inotify_simple